requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.14.2
bleach==6.2.0
blinker==1.9.0
boto3==1.40.52
botocore==1.40.52
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==45.0.7
deprecated==1.2.18
distro==1.9.0
fastuuid==0.13.5
feedparser==6.0.12
filelock==3.20.0
flask==3.1.2
flask-cors==6.0.1
flask-limiter==4.0.0
frozenlist==1.8.0
fsspec==2025.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.11
importlib-metadata==8.7.0
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kaggle==1.7.4.5
limits==5.6.0
litellm==1.78.0
markdown==3.9
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
multidict==6.7.0
numpy==2.3.3
openai==2.3.0
ordered-set==4.1.0
packaging==25.0
pandas==2.3.3
pilottai==0.2.7.37
propcache==0.4.1
protobuf==6.32.1
psutil==7.1.0
pycparser==2.23
pydantic==2.12.2
pydantic-core==2.41.4
pydantic-settings==2.11.0
pygments==2.19.2
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-http-client==3.3.7
python-slugify==8.0.4
pytz==2025.2
pyyaml==6.0.3
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
rich==14.2.0
rpds-py==0.27.1
s3transfer==0.14.0
schedule==1.2.2
scikit-learn==1.7.2
scipy==1.16.2
sendgrid==6.12.5
setuptools==80.9.0
sgmllib3k==1.0.0
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlitedict==2.1.0
substack-api==1.1.1
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
webencodings==0.5.1
werkzeug==3.1.3
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0